Update the zh/docs
