+++ app/zh/docs/layout.tsx
@@ -1,11 +1,23 @@
-'use client';
+import { getDocsNav, getLearnNav } from 'utils/docs/getDocProps';
+import DocsLayoutClient from './toc-layout-client';
 
-import { DocsNavigationProvider } from 'components/AppRouterMigrationComponents/Docs/DocsNavigationContext';
-
-export default function DocsLayout({
+export default async function DocsLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  return <DocsNavigationProvider>{children}</DocsNavigationProvider>;
+  // Fetch navigation data that will be shared across all docs pages
+  const [navDocData, navLearnData] = await Promise.all([
+    getDocsNav(),
+    getLearnNav(),
+  ]);
+
+  return (
+    <DocsLayoutClient
+      NavigationDocsData={navDocData}
+      NavigationLearnData={navLearnData}
+    >
+      {children}
+    </DocsLayoutClient>
+  );
 }
